requirements.txt
pandas
python-decouple
SQLAlchemy
PyMySQL
Flask
Flask-Cors
